requirements.txt
streamlit>=1.38.0
pymongo>=4.7.0
python-dotenv>=1.0.1
qrcode[pil]>=7.4.2
opencv-python-headless>=4.10.0.84
pillow>=10.3.0
numpy>=1.26.4
pandas>=2.2.2